Remove log translations
1. What is the problem?
According to the discussion on the Barcelona Design Summit[1], server
translated error messages are used for users to find solutions to their
deployment failures, as result the number of findings are less compared
with English errors.

2. What is the solution to the problem?
Log messages are no longer being translated. This removes all use of
the _LE, _LI, and _LW translation markers to simplify logging and to
avoid confusion with new contributions.

3. What features need to be implemented to the Tricircle
   to realize the solution?
None

[1] http://lists.openstack.org/pipermail/openstack-dev/2017-March/113383.html

Related-Bug: #1674567
Change-Id: I9ac745af5afc9f6f27884e763e0f1943dff603e6
Co-Authored-By: Victor Morales <victor.morales@intel.com>