Replace six.iteritems() with .items()
1.What is the problem?
As mentioned in [1], we should avoid using six.iteritems
to achieve iterators. In py2, the performance about list
should be negligible, please see the link [2].

2.What is the solution to the problem?
We can use dict.items instead, as it will return iterators
in PY3 as well and dict.items/keys will be more readable.

3.What the features need to be implemented to the Tricircle
to realize the solution?
None

[1] https://wiki.openstack.org/wiki/Python3
[2] http://lists.openstack.org/pipermail/openstack-dev/2015-June/066391.html

Change-Id: Ia564d0e5ec4b1bb63a747df75f282273c21867c7