Enforce to cleanup RST errors
PyPI is enforcing to not have syntax errors in package
descriptors. The README file is used as project's description.
According to the official documentation[1], PyPI supports three
different formats (plain text, reStructureText and Markdown).

This change changes the format of the README file to Markdown and
reports syntax errors in documentation. These errors need to be
fixed before a new release can be done.

[1] https://packaging.python.org/guides/making-a-pypi-friendly-readme/

Change-Id: Ia96882528df8a371cc8a89a7d9d74233c4d1b1de
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>