Change python3.5 job to python3.7 job on Stein+
This is a mechanically generated patch to add a unit test job running
under Python 3.7 and drop the Python 3.5 unit test job.

python3.5 was the only supported python3 version on Xenial. Now that
we have Bionic nodes supporting python3.6 and python3.7, let's switch
to testing with python3.7 in addition to python3.6 in Stein and
beyond.

See ML discussion here [1] for context.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-October/135626.html

Change-Id: Ic870e30988317611afd168c69fed84607df6a3e7
Story: #2004073
Task: #27458